pyproject.toml
[project]
name = "scrap-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.7.8",
    "fastapi[standard]>=0.128.0",
    "python-fasthtml>=0.12.37",
]
